// src/pages/create-agent/types.ts
export interface ToolResponse {
  _id: string;
  provider_id: string;
  type: string;
  form: {
    type: string;
    properties: Record<
      string,
      {
        type: string;
        description?: string;
      }
    >;
    required: string[];
  };
  meta_data: {
    schema: {
      info: {
        title: string;
        description: string;
      };
    };
  };
}
// ...
export interface ToolsProps {
  form: UseFormReturn<any, any, undefined>;
  onEnabledCountChange: (count: number) => void;
  apiTools: ToolResponse[];
  userTools: string[];
  loading: boolean;
  error: string | null;
  apiKey: string;
  setApiTools: (tools: ToolResponse[]) => void;
  setUserTools: (tools: string[]) => void;
}